# src/main/resources/api/data-repository-openapi.yaml
openapi: 3.0.1
info:
  title: Data Repository API
  description: >
    <details><summary>This document defines the REST API for the Terra Data Repository.</summary>
    <p>
    **Status: design in progress**

    There are a few top-level endpoints (besides some used by swagger):
     * / - generated by swagger: swagger API page that provides this
    documentation and a live UI for submitting REST requests
     * /status - provides the operational status of the service
     * /configuration - provides the basic configuration and information about the service
     * /api - is the authenticated and authorized Data Repository API
     * /ga4gh/drs/v1 - is a transcription of the Data Repository Service API

    The API endpoints are organized by interface. Each interface is separately versioned.
    <p>
    **Notes on Naming**
    <p>
    All of the reference items are suffixed with \"Model\".
    Those names are used as the class names in the generated Java code.
    It is helpful to distinguish these model classes from other related classes, like the DAO classes and the operation classes.
    </details>
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://app.terra.bio/#terms-of-service
  version: 0.1.0
servers:
  - url: /
security:
  - googleoauth:
      - openid
      - email
      - profile
  - oidc:
      - openid
      - email
      - profile
paths:
...
  /api/repository/v1/duos:
    get:
      tags:
        - duos
        - repository
      description: >
        Returns a list of all DUOS Firecloud groups managed by TDR.
        NOTE: This is an experimental feature and its response body may change.
      operationId: retrieveDuosFirecloudGroups
      responses:
        200:
          description: List of DUOS Firecloud groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DuosFirecloudGroupModel'
        500:
          description: >
            An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/repository/v1/duos/{duosId}:
    get:
      tags:
        - duos
        - repository
      description: >
        Returns a DUOS Firecloud group managed by TDR.
        NOTE: This is an experimental feature and its response body may change.
      operationId: retrieveDuosFirecloudGroup
      parameters:
        - name: duosId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DuosId'
      responses:
        200:
          description: DUOS Firecloud group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuosFirecloudGroupModel'
        404:
          description: >
            Not found - TDR has no record of of a DUOS dataset with the specified ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        500:
          description: >
            An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/repository/v1/duos/syncAuthorizedUsers:
    put:
      tags:
        - duos
        - repository
      description: >
        Sync the members of all TDR-managed DUOS Firecloud groups with the authorized users of their
        corresponding DUOS datasets.  Any snapshots linked to a DUOS dataset have its Firecloud
        group as a reader, which means its members are also snapshot readers.
        NOTE: This is an experimental feature and its response body may change.
      operationId: syncDuosDatasetsAuthorizedUsers
      responses:
        200:
          description: Sync complete - see contents to gauge overall success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuosFirecloudGroupsSyncResponse'
        403:
          description: >
            Unauthorized - the caller must be a TDR Admin or TDR-DUOS Integration Admin
            to facilitate sync.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        500:
          description: >
            An unexpected error occurred - Firecloud group contents may not have been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/repository/v1/duos/{duosId}/syncAuthorizedUsers:
    put:
      tags:
        - duos
        - repository
      description: >
        Sync the DUOS dataset's authorized users as members of their corresponding TDR-managed
        Firecloud group.  Any snapshots linked to the DUOS dataset have this Firecloud group
        as a reader, which means its members are also snapshot readers.
        NOTE: This is an experimental feature and its response body may change.
      operationId: syncDuosDatasetAuthorizedUsers
      parameters:
        - name: duosId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DuosId'
      responses:
        200:
          description: Firecloud group contents successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuosFirecloudGroupModel'
        403:
          description: >
            Unauthorized - the caller must be a TDR Admin or TDR-DUOS Integration Admin
            to facilitate sync.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: >
              Not found - TDR has no record of of a DUOS dataset with the specified ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        409:
          description: >
              Conflict - Firecloud group contents may have been updated, but TDR encountered a
              conflict when writing back to the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        500:
          description: >
              An unexpected error occurred - Firecloud group contents may not have been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
...
  /api/repository/v1/search/{id}/index:
    post:
      tags:
        - search
        - snapshots
        - repository
      description: Create a new search index for a snapshot
      operationId: createSearchIndex
      parameters:
        - name: id
          in: path
          description: A UUID to used to identify an object in the repository
          required: true
          schema:
            type: string
      requestBody:
        description: Index to create from SQL
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchIndexRequest'
        required: true
      responses:
        200:
          description: Return index result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchIndexModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to see snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - snapshot id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/repository/v1/search/query:
    post:
      tags:
        - search
        - snapshots
        - repository
      description: Run a query on the metadata of one or more snapshots
      operationId: querySearchIndices
      parameters:
        - name: offset
          in: query
          description: The number of results to skip before retrieving the next
            page
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: The number of search results to return
          schema:
            type: integer
            default: 1000
      requestBody:
        description: Search query to run against one or more snapshots
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchQueryRequest'
        required: true
      responses:
        200:
          description: List of search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchQueryResultModel'
        400:
          description: Bad request - one or more invalid ids, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to see one or more snapshots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - one or more snapshot ids do not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'

##############################################################################
## DRS STANDARD ENDPOINTS
##############################################################################
...
components:
  schemas:
...
    UniqueIdProperty:
      type: string
      format: uuid
      description: >
        Unique identifier for a dataset, snapshot, etc.
...
    ErrorModel:
      required:
        - message
      type: object
      properties:
        message:
          type: string
        errorDetail:
          type: array
          items:
            type: string
      description: >
        Representation of an error that occurred while processing a request.
        Some errors provide error details. For example, validation of input may report several problems.
        Each of those are reported in a separate error detail.
...
    DuosId:
      type: string
      description: DUOS dataset identifier
      example: DUOS-123456
...
    SearchIndexRequest:
      required:
        - sql
      type: object
      properties:
        sql:
          type: string
          description: >
            SQL to index BigQuery columns for searching
    SearchIndexModel:
      type: object
      properties:
        indexSummary:
          type: string
          description: Summary of this index for searching
...
    SearchQueryRequest:
      required:
        - query
      type: object
      properties:
        snapshotIds:
          type: array
          items:
            $ref: '#/components/schemas/UniqueIdProperty'
        query:
          type: string
          description: >
            Query to search against snapshot indices
    SearchQueryResultModel:
      type: object
      properties:
        result:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
          description: Results that match the search criteria
      description: >
        The available search results that match the filter criteria
...
    DuosFirecloudGroupModel:
      type: object
      description: >
        A Firecloud managed group for the authorized users of a DUOS dataset.
        This group has a TDR service account as its sole administrator.
        Only those DUOS datasets known to TDR will have Firecloud groups for their users.
      properties:
        id:
          $ref: '#/components/schemas/UniqueIdProperty'
        duosId:
          $ref: '#/components/schemas/DuosId'
        firecloudGroupName:
          description: Name of Firecloud managed group
          type: string
          example: DUOS-123456-users
        firecloudGroupEmail:
          description: Email of Firecloud managed group
          type: string
          example: DUOS-123456-users@dev.test.firecloud.org
        createdBy:
          description: The TDR service account which created this group
          type: string
          example: jade-k8-sa@broad-jade-dev.iam.gserviceaccount.com
        created:
          description: Time when TDR created this group
          type: string
        lastSynced:
          description: Last time TDR queried DUOS for its dataset's authorized users
          type: string
...
    DuosFirecloudGroupsSyncResponse:
      type: object
      description: >
        The outcome of a batch sync of all DUOS Firecloud groups.
      properties:
        synced:
          description: >
            The Firecloud groups whose contents were successfully updated.
          type: array
          items:
            $ref: '#/components/schemas/DuosFirecloudGroupModel'
        errors:
          description: >
            Errors which may have interfered in the syncing for select DUOS Firecloud groups.
          type: array
          items:
            $ref: '#/components/schemas/ErrorModel'

    ##############################################################################
    ## DRS STANDARD MODELS
    ##############################################################################
...
  securitySchemes:
    googleoauth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization
            https://www.googleapis.com/auth/cloud-billing.readonly: cloud billing authorization
    oidc:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          tokenUrl: /oauth2/token
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization